Unpin git-annex version
The previously pinned version (10.20221003) came with Git v2.30 -- which
is too old to support GIT_CONFIG_GLOBAL -- needed for the test setup.

This is a general problem with this datalad-installer, most variants
came with a bundled git -- and an ancient one.

In general there is little point in sticking with an old git-annex.
Rather we will have to ensure availability of a recent git-annex version
for deployment.