docs: draft user guide section on datalad access
This creates additional nesting level for user guide, with separate
pages for DataLad- and browser-based access, but I think it's worth
it.

The credentials section is a bit general now, but it is based on
something I wrote for DataLad-Redcap.

The clone URL pattern & its explanation is based on 9b66926 commit
message - when finalized, it's worth checking for up-to-datedness.