Chose a strategic title for the note on client side software
By including the term 'GDPR', anyone who visits the docs can find it
using the search bar. This way, we make it explicit that we are aware,
and might prevent arguments about GDPR compliance from the start.
By making the note an admonition, it not only gains a title, but also
a custom color highlight that makes it stand out on the page