Update docs requirements
Seeing that a build on Sphinx 7.0.1 succeeded, but a build on Sphinx
7.2.6 failed with Furo 2023.5.20 theme, I decided to bump furo version
and narrow down Sphinx to a minor release.

Furo changelog: https://pradyunsg.me/furo/changelog/

If you are reading this commit message and considering updating or
loosening the dependencies, then you probably should do it without
issues. I suppose even now an unpinned dependency would work; it's
probably on rare occasions when the released version of the theme
needs to catch up to Sphinx development, which likely caused us to
introduce pinning in first place. In either case, it would seem that
pinning Furo to a specific version, but Sphinx only to major release
is not a good idea.