docs: fix the singularity-calling script
The instructions for creating the shell script to simplify calling
singularity say "save it as icf-utils", but it seems that the provided
code block is a mixture of the file content (starts with a hashbang)
and a copy-paste snippet to create the file (ends with a redirection
to icf-utils).

This removes the seemingly incorrect redirection from the icf-utils
shell script definition.