Copy files (instead of git clone) on container build
This recipe would generate the container image from local files,
without having to go through github first.

The downside is that now the build process is sensitive to the working
directory when running singularity build - it needs to be the
repositository base dir.

This potentially removes the informal role of the GitHub main branch
as the "rolling release" of sources for the image. Instead, building
files locally becomes easier.